hook file is generated with init
