+++ packages/houdini/cmd/utils/index.ts
@@ -3,3 +3,4 @@ export * from './commonjs'
 export * from './writeFile'
 export * from './graphql'
 export * from './murmur'
+export * from './writeSchema'
